backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.0.8
chromadb==0.5.4
sentence-transformers==3.0.1
pydantic==2.5.3
pandas==2.2.2
openpyxl==3.1.4
python-docx==1.1.2
pypdf==4.2.0
python-dotenv==1.0.0